requirements.txt
streamlit>=1.31.0
httpx>=0.27.0
pypdf>=4.0.0
python-docx>=1.1.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
pytest>=8.0.0
ruff>=0.3.0
